Actually test for implicit conversion.
Note that tests in `SimpleSFXDelegateSpec` are named the same, but do not actually test that required implicit conversions exists. Testing for implicit conversion in a generic way is difficult since they are resolved at compile time. The approach here relies on a user of the actually implement those test using helper method `canConvert`. That method is designed to force compiler into providing a handle to actual implicit conversion function (if one exists). `SimpleSFXDelegateSpec`requires user to provide conversions but does not require that those conversions are implicit.

--HG--
branch : enum_experiments