Remove duplicate code. There were anout 19 implicit conversions for event handllers that can be replaced by a single paramtrized one: scalafx.event.EventIncludes.eventClosureWrapperWithParam.
Also instead of using `Any` as a closure return type, using a type parameter `R`, similar to the way it is done in Scala collections library.
Somehow, two of the issues illustrated in a demo for Issue 102 were fixed: they no longer cmpile, which is a correct behaviour, so corresponding code in IncompleteClickHandler was commented out.